Refactor code to seperate api code from redux
