+++ src/components/NewBook.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
-import { postBook } from '../redux/books/books';
+import { postBook } from '../redux/api';
 
 function NewBook() {
   const dispatch = useDispatch();
